==> final_code/1_data_wrangling.ipynb <==
import pandas as pd
import numpy as np
from molvs import standardize_smiles

### Generate train data and labels (without standardizing)

# read tox21 train dataset with labels
# selected nr_ahr assay for final project
nr_ahr = pd.read_csv('../raw_data/nr-ahr.smiles', sep='\t', header=None)
nr_ahr = nr_ahr.rename(columns={0: 'compounds', 1: 'id', 2: 'label'})
nr_ahr.head()

# check train dataset for duplicates
print('Number of compounds in raw dataset:', len(nr_ahr['compounds']))
print('Number of unique compounds:', len(nr_ahr['compounds'].unique()))

# remove duplicates
nr_ahr = nr_ahr.drop_duplicates(subset='compounds').reset_index(drop=True)
len(nr_ahr)

# check label column for inconclusive or tests not run (x)
nr_ahr.label.unique()

# generate duplicate-free training dataset csv file
#nr_ahr.to_csv('../processed_data/nr_ahr_train.csv')

# split into train_data and train_labels
train_data = nr_ahr['compounds']
train_labels = nr_ahr['label']

print(len(train_data))
print(len(train_labels))

### Standardize train data using MolVS

# identify rows with SMILES errors
train_error_rows = []

for i, row in enumerate(nr_ahr.compounds):
    try:
        standardize_smiles(row)
    except Exception as e: 
        print('Error at index {}: {!r}'.format(i, row))
        print(e)
        train_error_rows.append(i)

print()
print('Number of rows with errors:', len(train_error_rows))

# add error rows that were missed by for loop
train_error_rows.append(2888)
train_error_rows.append(4194)

# remove rows with SMILES errors
train_drop_rows = [nr_ahr.index[i] for i in train_error_rows]
print(len(train_drop_rows))

nr_ahr_std = nr_ahr.drop(train_drop_rows).reset_index(drop=True)
print(len(nr_ahr_std))

# standardize compound SMILES using MolVS
nr_ahr_std['std_compounds'] = nr_ahr_std.compounds.map(lambda x: standardize_smiles(x))
print(len(nr_ahr_std))
nr_ahr_std.head()

# generate standardized training dataset csv file
#nr_ahr_std.to_csv('../processed_data/nr_ahr_std_train.csv')

# split into train_data and train_labels
train_data_std = nr_ahr_std['compounds']
train_labels_std = nr_ahr_std['label']

print(len(train_data_std))
print(len(train_labels_std))

### Generate test data and labels (without standardizing)

# read tox21 score dataset
score = pd.read_csv('../raw_data/tox21_10k_challenge_score.smiles', sep='\t')
score = score.rename(columns={'#SMILES': 'compounds', 'Sample ID': 'id'})
score.head()

# read tox21 score labels
score_results = pd.read_csv('../raw_data/tox21_10k_challenge_score.txt', sep='\t')
score_results.head()

# pick out NR-AhR labels
score_results = score_results[['Sample ID', 'NR-AhR']]
score_results = score_results.rename(columns={'Sample ID': 'id', 'NR-AhR': 'label'})
score_results.head()

# merge score dataset with NR-AhR labels using compound ID
nr_ahr_test = score.merge(score_results, on='id')
nr_ahr_test.head()

# check test dataset for duplicates
print('Number of compounds in raw dataset:', len(nr_ahr_test['compounds']))
print('Number of unique compounds:', len(nr_ahr_test['compounds'].unique()))

# remove duplicates
nr_ahr_test = nr_ahr_test.drop_duplicates(subset='compounds').reset_index(drop=True)
len(nr_ahr_test)

# check label column for inconclusive/tests not run (x)
nr_ahr_test.label.unique()

# remove rows with inconclusive/tests not run (x)
nr_ahr_test = nr_ahr_test[nr_ahr_test.label != 'x']
print(nr_ahr_test.label.unique())
print(len(nr_ahr_test))

# generate duplicate-free training dataset csv file
#nr_ahr_test.to_csv('../processed_data/nr_ahr_test.csv')

# split into test_data and test_labels
test_data = nr_ahr_test['compounds']
test_labels = nr_ahr_test['label']

print(len(test_data))
print(len(test_labels))

### Standardize test data using MolVS

# identify rows with SMILES errors
test_error_rows = []

for i, row in enumerate(nr_ahr_test.compounds):
    try:
        standardize_smiles(row)
    except Exception as e: 
        print('Error at index {}: {!r}'.format(i, row))
        print(e)
        train_error_rows.append(i)

print()
print('Number of rows with errors:', len(test_error_rows))

# add error rows that were missed by the for loop
test_error_rows = [303, 344]

# remove rows with SMILES errors
test_drop_rows = [nr_ahr_test.index[i] for i in test_error_rows]
print(len(test_drop_rows))

nr_ahr_test_std = nr_ahr_test.drop(test_drop_rows).reset_index(drop=True)

print(len(nr_ahr_test_std))

# standardize compound SMILES using MolVS
nr_ahr_test_std['std_compounds'] = nr_ahr_test_std.compounds.map(lambda x: standardize_smiles(x))
print(len(nr_ahr_test_std))
nr_ahr_test_std.head()

# generate duplicate-free training dataset csv file
#nr_ahr_test_std.to_csv('../processed_data/nr_ahr_test_std.csv')

# split into test_data and test_labels
test_data_std = nr_ahr_test['compounds']
test_labels_std = nr_ahr_test['label']

print(len(test_data))
print(len(test_labels))
